1 - Erro 500 - ao consultar ordem de serviço 2 - Colocar a data na ordem de pagamento como not null e a dataPagamento permitindo em branco (está ao contrario) 3 - Inserir uns registros de ordem de pagamento default 4 - Erro 500 - ao consultar ordem de pagamento

+++ src/main/java/br/com/fatesg/fabrica/projetofabrica/controlador/OrdemPagamentoResource.java
@@ -34,7 +34,7 @@ public class OrdemPagamentoResource {
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<OrdemPagamento> buscar(@PathVariable int id) {
+	public ResponseEntity<OrdemPagamento> buscar(@PathVariable long id) {
 		OrdemPagamento obj = negocio.findById(id);
 		
 		if (obj == null) {
